requirements.txt
asgiref==3.2.10
boto3==1.16.4
botocore==1.19.4
Django==3.1
django-cors-headers==3.5.0
django-storages==1.10.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
jmespath==0.10.0
numpy==1.19.2
pandas==1.1.2
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
